First version of ResFileUtil
This implements more or less the same functionality as the old ATL/COM C++ project MLResUtil, with more or less the same API and more or less the same algorithms.
The test project doesn't deserve to be called a test project. It is just a framework for me to debug the functions. It also refers to local files on my PC.